requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
libvirt-python>=9.0.0
pefile>=2023.2.7
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
